Rework implementation to use composition instead of inheritance.
