+++ processing/src/main/java/org/apache/druid/query/aggregation/BufferAggregator.java
@@ -38,14 +38,21 @@ import java.nio.ByteBuffer;
  * @see VectorAggregator, the vectorized version
  */
 @ExtensionPoint
-public interface BufferAggregator extends BaseBufferAggregator, HotLoopCallee
+public interface BufferAggregator extends HotLoopCallee
 {
   /**
-   * {@inheritDoc}
+   * Initializes the buffer location
    *
-   * Overridden because this method is a {@link HotLoopCallee} and the superinterface method is not.
+   * Implementations of this method must initialize the byte buffer at the given position
+   *
+   * <b>Implementations must not change the position, limit or mark of the given buffer</b>
+   *
+   * This method must not exceed the number of bytes returned by {@link AggregatorFactory#getMaxIntermediateSizeWithNulls}
+   * in the corresponding {@link AggregatorFactory}
+   *
+   * @param buf byte buffer to initialize
+   * @param position offset within the byte buffer for initialization
    */
-  @Override
   @CalledFromHotLoop
   void init(ByteBuffer buf, int position);
 
@@ -57,20 +64,33 @@ public interface BufferAggregator extends BaseBufferAggregator, HotLoopCallee
    *
    * <b>Implementations must not change the position, limit or mark of the given buffer</b>
    *
-   * @param buf      byte buffer storing the byte array representation of the aggregate
+   * @param buf byte buffer storing the byte array representation of the aggregate
    * @param position offset within the byte buffer at which the current aggregate value is stored
    */
   @CalledFromHotLoop
   void aggregate(ByteBuffer buf, int position);
 
   /**
-   * {@inheritDoc}
+   * Returns the intermediate object representation of the given aggregate.
    *
-   * Overridden because this method is a {@link HotLoopCallee} and the superinterface method is not.
+   * Converts the given byte buffer representation into an intermediate aggregate Object
+   *
+   * <b>Implementations must not change the position, limit or mark of the given buffer.</b>
+   *
+   * <b>
+   * The object returned must not have any references to the given buffer (i.e., make a copy), since the
+   * underlying buffer is a shared resource and may be given to another processing thread
+   * while the objects returned by this aggregator are still in use.
+   * </b>
+   *
+   * <b>If the corresponding {@link AggregatorFactory#combine(Object, Object)} method for this aggregator
+   * expects its inputs to be mutable, then the object returned by this method must be mutable.</b>
+   *
+   * @param buf byte buffer storing the byte array representation of the aggregate
+   * @param position offset within the byte buffer at which the aggregate value is stored
+   * @return the Object representation of the aggregate
    */
-  @Override
   @Nullable
-  @CalledFromHotLoop
   Object get(ByteBuffer buf, int position);
 
   /**
@@ -84,9 +104,8 @@ public interface BufferAggregator extends BaseBufferAggregator, HotLoopCallee
    * have an {@link AggregatorFactory#getType()} ()} of {@link org.apache.druid.segment.column.ValueType#FLOAT}.
    * If unimplemented, throwing an {@link UnsupportedOperationException} is common and recommended.
    *
-   * @param buf      byte buffer storing the byte array representation of the aggregate
+   * @param buf byte buffer storing the byte array representation of the aggregate
    * @param position offset within the byte buffer at which the aggregate value is stored
-   *
    * @return the float representation of the aggregate
    */
   float getFloat(ByteBuffer buf, int position);
@@ -102,9 +121,8 @@ public interface BufferAggregator extends BaseBufferAggregator, HotLoopCallee
    * have an {@link AggregatorFactory#getType()} of  of {@link org.apache.druid.segment.column.ValueType#LONG}.
    * If unimplemented, throwing an {@link UnsupportedOperationException} is common and recommended.
    *
-   * @param buf      byte buffer storing the byte array representation of the aggregate
+   * @param buf byte buffer storing the byte array representation of the aggregate
    * @param position offset within the byte buffer at which the aggregate value is stored
-   *
    * @return the long representation of the aggregate
    */
   long getLong(ByteBuffer buf, int position);
@@ -124,15 +142,19 @@ public interface BufferAggregator extends BaseBufferAggregator, HotLoopCallee
    * This default method is added to enable smooth backward compatibility, please re-implement it if your aggregators
    * work with numeric double columns.
    *
-   * @param buf      byte buffer storing the byte array representation of the aggregate
+   * @param buf byte buffer storing the byte array representation of the aggregate
    * @param position offset within the byte buffer at which the aggregate value is stored
-   *
    * @return the double representation of the aggregate
    */
   default double getDouble(ByteBuffer buf, int position)
   {
     return (double) getFloat(buf, position);
   }
+
+  /**
+   * Release any resources used by the aggregator
+   */
+  void close();
 
   /**
    * {@inheritDoc}
@@ -144,6 +166,29 @@ public interface BufferAggregator extends BaseBufferAggregator, HotLoopCallee
   default void inspectRuntimeShape(RuntimeShapeInspector inspector)
   {
     // nothing to inspect
+  }
+
+  /**
+   * Relocates any cached objects.
+   * If underlying ByteBuffer used for aggregation buffer relocates to a new ByteBuffer, positional caches(if any)
+   * built on top of old ByteBuffer can not be used for further {@link BufferAggregator#aggregate(ByteBuffer, int)}
+   * calls. This method tells the BufferAggregator that the cached objects at a certain location has been relocated to
+   * a different location.
+   *
+   * Only used if there is any positional caches/objects in the BufferAggregator implementation.
+   *
+   * If relocate happens to be across multiple new ByteBuffers (say n ByteBuffers), this method should be called
+   * multiple times(n times) given all the new positions/old positions should exist in newBuffer/OldBuffer.
+   *
+   * <b>Implementations must not change the position, limit or mark of the given buffer</b>
+   *
+   * @param oldPosition old position of a cached object before aggregation buffer relocates to a new ByteBuffer.
+   * @param newPosition new position of a cached object after aggregation buffer relocates to a new ByteBuffer.
+   * @param oldBuffer old aggregation buffer.
+   * @param newBuffer new aggregation buffer.
+   */
+  default void relocate(int oldPosition, int newPosition, ByteBuffer oldBuffer, ByteBuffer newBuffer)
+  {
   }
 
   /**
@@ -163,4 +208,5 @@ public interface BufferAggregator extends BaseBufferAggregator, HotLoopCallee
   {
     return false;
   }
+
 }
